refactor: restructure ConsiderationProposal types and interfaces
